% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem.R
\name{cycle_vector_to_length}
\alias{cycle_vector_to_length}
\title{Converts a vector to a particular length by cycling it.}
\usage{
cycle_vector_to_length(v, len)
}
\arguments{
\item{v}{vector. The vector to cycle.}

\item{len}{integer. The length to make it.}
}
\description{
Converts a vector to a particular length by cycling it.
}
