% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem-package.R
\docType{package}
\name{lorem-package}
\alias{lorem}
\alias{lorem-package}
\title{lorem: Generate Lorem Ipsum Text}
\description{
Quickly generate lorem ipsum placeholder text. Easy to integrate in RMarkdown documents. Includes an RStudio addin to insert lorem ipsum into the current document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/lorem}
  \item \url{http://pkg.garrickadenbuie.com/lorem/}
  \item Report bugs at \url{https://github.com/gadenbuie/lorem/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
