% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem.R
\name{lorem}
\alias{lorem}
\title{Print lorem ipsum to screen}
\usage{
lorem(x = 1)
}
\arguments{
\item{x}{integer Number of chapters to print
'x' has to be an integer >= 0 (default = 1)}
}
\description{
\code{lorem} prints 'x' chapters of lorem ipsum dummy text to the screen.
}
\examples{
lorem(1)
}
