% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem.R
\name{words}
\alias{words}
\title{Convert a string into a vector, word separated.}
\usage{
words(str)
}
\arguments{
\item{str.}{The character to convert to words.}
}
\description{
Convert a string into a vector, word separated.
}
