% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lorem.R
\name{lorem}
\alias{lorem}
\title{Lorem Ipsum Generator}
\usage{
lorem(n)
}
\arguments{
\item{n}{integer. The amount of lorem words to generate.}
}
\value{
glorius lorem.
}
\description{
Lorem Ipsum Generator
}

